+++ resources/views/pages/dashboard/_algolia-indices.blade.php
@@ -1,0 +1,24 @@
+<div class="col-sm-12 col-md-3">
+    <div class="card">
+        <h2 class="card-title">{{ $algolia['products'] }}</h2>
+        <p class="card-content">{{ __('Records in products') }}</p>
+    </div>
+</div>
+<div class="col-sm-12 col-md-3">
+    <div class="card">
+        <h2 class="card-title">{{ $algolia['categories'] }}</h2>
+        <p class="card-content">{{ __('Records in categories') }}</p>
+    </div>
+</div>
+<div class="col-sm-12 col-md-3">
+    <div class="card">
+        <h2 class="card-title">{{ $algolia['brands'] }}</h2>
+        <p class="card-content">{{ __('Records in brands') }}</p>
+    </div>
+</div>
+<div class="col-sm-12 col-md-3">
+    <div class="card">
+        <h2 class="card-title">{{ $algolia['users'] }}</h2>
+        <p class="card-content">{{ __('Records in users') }}</p>
+    </div>
+</div>