+++ resources/views/pages/orders/index.blade.php
@@ -25,7 +25,7 @@
                             <th>{{ __('Email') }}</th>
                             <th>{{ __('Name') }}</th>
                             <th>{{ __('Phone') }}</th>
-                            <th>{{ __('Total Price') }}</th>
+                            <th>{{ __('Price') }}</th>
                         </tr>
                     </thead>
                     <tbody>
@@ -42,7 +42,7 @@
                                 <td>{{ $record->getUser('email') }}</td>
                                 <td>{{ $record->getUser() }}</td>
                                 <td>{{ $record->getUser('phone') }}</td>
-                                <td>{{ $record->total_price }}</td>
+                                <td>{{ shopperMoney($record->total_price, setting('site_currency')) }}</td>
                             </tr>
                         @empty
                             <tr>