+++ changelog.md
@@ -2,14 +2,41 @@
 
 All notable changes to `Shopper` will be documented in this file.
 
-## Version 1.0
+## Version 1.0.2
+### Added
+- Twitter and Facebook card on home dashboard
+- `thujohn/twitter` to manage twitter 
+- Shopper version status 
 
+### Changed
+- ShopperServiceProvider (register new config file `ttwitter.php`)
+
+## Version 1.0.1
+### Added
+- Algolia indexes informations to dashboard
+- Laravel Viewable model migration
+
+## Version 1.0
+### Added
+- Laravel-mail-editor
+- Documentation
+- Coupon Management
+- Discount Management
+
+### Updated
+- Shopper installation command
+
+### Fixed
+- Error migration on order management
+
+## Version 1.0.2-beta
+### Added
+- Filemanager
+
+## Version 1.0.1-beta 
+### Remove
+-  Creation of admin user during shopper installation
+
+## Version 1.0.0-beta 
 ### Added
 - Everything
-
-### Version 1.0.2-beta
-- Add Filemanager
-
-### Version 1.0.0
-- Add Laravel-mail-editor
-- Add documentation